src/main/java/net/bishnu/leetcode/validParentheses/ValidParentheses.java
package net.bishnu.leetcode.validParentheses;

import java.util.ArrayDeque;
import java.util.Deque;
import java.util.Stack;

/**
 * 20. Valid Parentheses
 * https://leetcode.com/problems/valid-parentheses/discuss/
 */
public class ValidParentheses{
    public boolean isValid(String s) {
        Deque<Character> stack = new ArrayDeque<>();
        for (char c : s.toCharArray()) {
            if (c == '(')
                stack.push(')');
            else if (c == '{')
                stack.push('}');
            else if (c == '[')
                stack.push(']');
            else if (stack.isEmpty() || stack.pop() != c)
                return false;
        }
        return stack.isEmpty();
    }
}
